This is a basic splitting of TorchToLinalg.cpp.
Before this file gets even larger let's split it.

1. I am yet to write all the common function under a namespace.
3. This PR is just for initial disucssion on splitting.
