+++ lib/Conversion/TorchToLinalg/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_mlir_conversion_library(TorchMLIRTorchToLinalg
   TorchToLinalg.cpp
+  CommonTorchToLinalg.cpp
+  Math/Matmul.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${PROJECT_SOURCE_DIR}/include/torch-mlir/Conversion/TorchToLinalg
