+++ lib/Conversion/TorchToLinalg/CommonTorchToLinalg.cpp
@@ -1,0 +1,243 @@
+#include "torch-mlir/Conversion/TorchToLinalg/CommonTorchToLinalg.h"
+
+using namespace mlir;
+using namespace mlir::torch;
+using namespace mlir::torch::Torch;
+
+LogicalResult verifyLinalgCompatibleTypes(Operation *op,
+                                          PatternRewriter &rewriter) {
+  // Check the value tensor is ranked as expected by Linalg.
+  // TODO: Remove this check but use a separate verification pass to verify the
+  // invariants expected by later passes.
+  auto isValidLinalgType = [](Type type) {
+    auto tensor = type.dyn_cast<ValueTensorType>();
+    return !tensor ||
+           tensor.toBuiltinTensor().dyn_cast_or_null<RankedTensorType>();
+  };
+
+  bool valid = llvm::all_of(op->getOperandTypes(), isValidLinalgType) &&
+               llvm::all_of(op->getResultTypes(), isValidLinalgType);
+  if (!valid)
+    return rewriter.notifyMatchFailure(op, "type cannot be lowered to linalg");
+  return success();
+}
+
+LogicalResult checkNotNone(PatternRewriter &rewriter, Operation *op, Value v) {
+  Type type = v.getType();
+  if (type.isa<OptionalType>() || type.isa<Torch::NoneType>() ||
+      type.isa<mlir::NoneType>())
+    return rewriter.notifyMatchFailure(op, "unimplemented None type arg");
+  return success();
+}
+
+// Generate IR: dim = dim >= 0 ? dim : dim + inputRank
+Value toPositiveDimDynamic(OpBuilder &b, Location loc, Value dim,
+                           Value inputRank) {
+  assert(dim.getType().isa<IntegerType>() &&
+         "dim arg of toPositiveDim must be integer type");
+  Value dimAddInputRank = b.create<arith::AddIOp>(loc, dim, inputRank);
+  Value cst0 =
+      b.create<arith::ConstantOp>(loc, b.getZeroAttr(inputRank.getType()));
+  Value predDimGEZero =
+      b.create<arith::CmpIOp>(loc, arith::CmpIPredicate::sge, dim, cst0);
+  Value dimInt = b.create<SelectOp>(loc, predDimGEZero, dim, dimAddInputRank);
+  return dimInt;
+}
+
+// Generate IR: assert(dim >= 0 && dim < inputRank)
+void assertIsValidDim(OpBuilder &b, Location loc, Value dim, Value inputRank) {
+  assert(dim.getType().isa<IntegerType>() &&
+         "dim arg of assertIsValidDim must be integer type");
+  Value cst0 =
+      b.create<arith::ConstantOp>(loc, b.getZeroAttr(inputRank.getType()));
+  Value predGEZero =
+      b.create<arith::CmpIOp>(loc, arith::CmpIPredicate::sge, dim, cst0);
+  b.create<AssertOp>(loc, predGEZero,
+                     b.getStringAttr("dim must be greater or equal to zero"));
+  Value predLTInputRank =
+      b.create<arith::CmpIOp>(loc, arith::CmpIPredicate::slt, dim, inputRank);
+  b.create<AssertOp>(loc, predLTInputRank,
+                     b.getStringAttr("dim must be smaller than inputRank"));
+}
+
+// Hack to deal with the Torch list type arguments which is not supported end
+// to end. Constant values can be be extracted directly and non constant
+// list values are not supported.
+// TODO: loose this constraint when properly support list type
+bool isConstantIntListMatching(Value value, SmallVectorImpl<int64_t> &expects) {
+  SmallVector<int64_t> intValues;
+  if (!matchPattern(value, m_TorchConstantIntList(intValues)))
+    return false;
+
+  if (intValues.size() != expects.size())
+    return false;
+
+  for (auto it : llvm::zip(intValues, expects)) {
+    if (std::get<0>(it) != std::get<1>(it))
+      return false;
+  }
+  return true;
+}
+
+Value castIntToIndex(OpBuilder &b, Location loc, Value v) {
+  assert(v.getType().isa<IntegerType>() && "must be called with integer type");
+  return b.create<arith::IndexCastOp>(loc, b.getIndexType(), v);
+}
+
+Value castIndexToInt(OpBuilder &b, Location loc, Value idx) {
+  assert(idx.getType().isa<IndexType>() && "must be called with integer type");
+  return b.create<arith::IndexCastOp>(loc, b.getI64Type(), idx);
+}
+
+Value getDimOp(OpBuilder &b, Location loc, Value v, int dimension) {
+  return b.create<tensor::DimOp>(loc, v, dimension);
+}
+
+void checkDimEqualHelper(OpBuilder &b, Location loc, Value lhsDim,
+                         Value rhsDim) {
+  Type lhsType = lhsDim.getType();
+  Type rhsType = rhsDim.getType();
+  auto checkIntOrIndex = [](Type type) {
+    assert(type.isa<IntegerType>() ||
+           type.isa<IndexType>() && "must be either integer or index type");
+  };
+  checkIntOrIndex(lhsType);
+  checkIntOrIndex(rhsType);
+  Value lhsDimInt = lhsType.isIndex() ? castIndexToInt(b, loc, lhsDim) : lhsDim;
+  Value rhsDimInt = rhsType.isIndex() ? castIndexToInt(b, loc, rhsDim) : rhsDim;
+  Value contractingDimEqual = b.create<arith::CmpIOp>(
+      loc, arith::CmpIPredicate::eq, lhsDimInt, rhsDimInt);
+  b.create<AssertOp>(loc, contractingDimEqual,
+                     b.getStringAttr("mismatching contracting dimension"));
+}
+
+SmallVector<Value> getTensorSizesUntilDim(OpBuilder &b, Location loc,
+                                          Value tensor, int dim) {
+  RankedTensorType type = tensor.getType().cast<RankedTensorType>();
+  assert(dim < type.getRank() &&
+         "The given dim must be smaller than tensor rank");
+  (void)type;
+  SmallVector<Value> sizes;
+  for (int i = 0; i <= dim; i++)
+    sizes.push_back(getDimOp(b, loc, tensor, i));
+  return sizes;
+}
+
+SmallVector<Value> getTensorSizes(OpBuilder &b, Location loc, Value tensor) {
+  RankedTensorType type = tensor.getType().cast<RankedTensorType>();
+  return getTensorSizesUntilDim(b, loc, tensor, type.getRank() - 1);
+}
+
+Value createZeroInitTensor(OpBuilder &b, Location loc, ValueRange sizes,
+                           Type elemTy) {
+  Value initTensor = b.create<linalg::InitTensorOp>(loc, sizes, elemTy);
+  RankedTensorType type = initTensor.getType().cast<RankedTensorType>();
+  Value c0 =
+      b.create<arith::ConstantOp>(loc, b.getZeroAttr(type.getElementType()));
+  return b.create<linalg::FillOp>(loc, c0, initTensor).getResult(0);
+}
+
+// Helper function to caculate the output tensor dims for convolution-like ops.
+// Along each dim:
+// dim_out =
+//  floor((dim_in + 2 * padding - dilation * (kernelSize - 1) - 1) / stride) + 1
+Value getOutputDimForConvOps(OpBuilder &b, Location loc, Value in,
+                             Value paddingInt, Value dilationInt,
+                             Value kernelSizeInt, Value strideInt) {
+  Value c1 = b.create<arith::ConstantOp>(loc, b.getI64IntegerAttr(1));
+  Value c2 = b.create<arith::ConstantOp>(loc, b.getI64IntegerAttr(2));
+
+  Value doublePadding = b.create<arith::MulIOp>(loc, paddingInt, c2);
+  // in + 2 * padding
+  Value inAddDoublePadding =
+      b.create<arith::AddIOp>(loc, castIndexToInt(b, loc, in), doublePadding);
+
+  // dilation * (kernelSize - 1)
+  Value kernelSizeSub1 = b.create<arith::SubIOp>(loc, kernelSizeInt, c1);
+  Value dilationTimesKernelSize =
+      b.create<arith::MulIOp>(loc, dilationInt, kernelSizeSub1);
+
+  Value temp =
+      b.create<arith::SubIOp>(loc, inAddDoublePadding, dilationTimesKernelSize);
+  Value dividend = b.create<arith::SubIOp>(loc, temp, c1);
+  Value division = b.create<arith::FloorDivSIOp>(loc, dividend, strideInt);
+  Value out = b.create<arith::AddIOp>(loc, division, c1);
+  return castIntToIndex(b, loc, out);
+}
+
+SmallVector<Value> getAsConstantIntValues(OpBuilder &b, Location loc,
+                                          SmallVectorImpl<int64_t> &ints) {
+  return llvm::to_vector<4>(llvm::map_range(ints, [&](int64_t val) -> Value {
+    return b.create<arith::ConstantOp>(loc,
+                                       b.getIntegerAttr(b.getI64Type(), val));
+  }));
+}
+
+SmallVector<Value> getAsConstantIndexValues(OpBuilder &b, Location loc,
+                                            SmallVectorImpl<int64_t> &ints) {
+  return llvm::to_vector<4>(llvm::map_range(ints, [&](int64_t val) -> Value {
+    return b.create<arith::ConstantOp>(loc, b.getIndexAttr(val));
+  }));
+}
+
+SmallVector<OpFoldResult> getAsOpFoldResult(OpBuilder &b, Location loc,
+                                            SmallVectorImpl<int64_t> &ints) {
+  return llvm::to_vector<4>(llvm::map_range(
+      ints, [&](int64_t val) -> OpFoldResult { return b.getIndexAttr(val); }));
+}
+
+// This is a temporary solution to deal with types that are not fully supported
+// like list, dict. For those container tyes, this helper can be used to
+// convert their elements to valid target type.
+// TODO: remove this when list gets full support.
+SmallVector<Value> getTypeConvertedValues(OpBuilder &b, Location loc,
+                                          TypeConverter *converter,
+                                          SmallVectorImpl<Value> &vs) {
+  return llvm::to_vector<4>(llvm::map_range(vs, [&](Value v) {
+    return converter->materializeTargetConversion(
+        b, loc, converter->convertType(v.getType()), v);
+  }));
+}
+
+// Helper function to get the padding tensor given the padding int values.
+// It's assumed that the padding on the low end and high end are the same.
+Value getPaddedTensor(Operation *op, OpBuilder &b, Value &input,
+                      SmallVectorImpl<int64_t> &paddingInts) {
+  assert(input.getType().isa<RankedTensorType>() &&
+         "input must be RankedTensorType");
+  Location loc = op->getLoc();
+  Value c0 = b.create<arith::ConstantOp>(
+      loc,
+      b.getZeroAttr(input.getType().cast<RankedTensorType>().getElementType()));
+  SmallVector<OpFoldResult> paddings = getAsOpFoldResult(b, loc, paddingInts);
+  Type ranked4DTensorType = linalg::PadTensorOp::inferResultType(
+      input.getType().cast<RankedTensorType>(), paddingInts, paddingInts);
+  Value paddedInput = linalg::PadTensorOp::createPadScalarOp(
+      ranked4DTensorType, input, c0, /*low=*/paddings, /*high=*/paddings,
+      /*packing=*/false, loc, b);
+  return paddedInput;
+}
+
+Value buildNormalCdf(OpBuilder &b, Location &loc, Value x, Value mean,
+                     Value sigma) {
+  Type elementType = x.getType();
+  Value xMinusMean = b.create<arith::SubFOp>(loc, x, mean);
+  Value two = b.create<arith::ConstantOp>(loc, FloatAttr::get(elementType, 2));
+  Value sqrt2 = b.create<math::SqrtOp>(loc, two);
+  Value erfArg = b.create<arith::DivFOp>(loc, xMinusMean, sqrt2);
+  Value erf = b.create<math::ErfOp>(loc, erfArg);
+  Value one = b.create<arith::ConstantOp>(loc, FloatAttr::get(elementType, 1));
+  Value erfPlus1 = b.create<arith::AddFOp>(loc, one, erf);
+  Value oneHalf =
+      b.create<arith::ConstantOp>(loc, FloatAttr::get(elementType, 0.5));
+  Value normalCdf = b.create<arith::MulFOp>(loc, oneHalf, erfPlus1);
+  return normalCdf;
+}
+
+Value buildUnitNormalCdf(OpBuilder &b, Location &loc, Value x) {
+  Type elementType = x.getType();
+  Value zero = b.create<arith::ConstantOp>(loc, FloatAttr::get(elementType, 0));
+  Value one = b.create<arith::ConstantOp>(loc, FloatAttr::get(elementType, 1));
+  return buildNormalCdf(b, loc, x, zero, one);
+}
+
